Consolidated Add and Edit into a single view/viewmodel
